Allow for streams to be "sealed".
A sealed stream is a stream which will never provide any more events. The
notion of sealing allows consumers to permanently stop attempting to consume
from a stream, and hence allows the projector to cleanup resource versions
related to the sealed stream (though this latter part has not yet been
implemented).